+++ xLights/outputs/xxxEthernetOutput.cpp
@@ -17,7 +17,7 @@
 #include "../UtilFunctions.h"
 #include "../utils/ip_utils.h"
 
-#include "spdlog/spdlog.h"
+#include <log.h>
 
 static const int32_t xxxCHANNELSPERPACKET = 1200;
 
@@ -61,7 +61,7 @@ void xxxEthernetOutput::Heartbeat(int mode, const std::string& localIP) {
             if (!__datagram->IsOk() || __datagram->Error()) {
                 spdlog::error("xxxEthernetOutput: {} Error creating xxxEthernet heartbeat datagram => {} : {}.",
                     (const char*)localaddr.IPAddress().c_str(),
-                    __datagram->LastError(),
+                    (int)__datagram->LastError(),
                     (const char*)DecodeIPError(__datagram->LastError()).c_str());
                 delete __datagram;
                 __datagram = nullptr;
@@ -105,7 +105,7 @@ void xxxEthernetOutput::OpenDatagram() {
         _datagram = nullptr;
     }
     else if (_datagram->Error()) {
-        spdlog::error("xxxEthernetOutput: {} Error creating xxxEthernet datagram => {} : {}.", (const char*)localaddr.IPAddress().c_str(), _datagram->LastError(), (const char*)DecodeIPError(_datagram->LastError()).c_str());
+        spdlog::error("xxxEthernetOutput: {} Error creating xxxEthernet datagram => {} : {}.", (const char*)localaddr.IPAddress().c_str(), (int)_datagram->LastError(), (const char*)DecodeIPError(_datagram->LastError()).c_str());
         delete _datagram;
         _datagram = nullptr;
     }
